Move PMAccPrinter to separate file
